Move client to the right spot
