+++ azuredevops/internal/service/securityroles/data_securityrole_definitions.go
@@ -7,8 +7,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/client"
-	securityroles "github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils/clients"
 	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils/tfhelper"
+	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/utils/securityroles"
 )
 
 func DataSecurityRoleDefinitions() *schema.Resource {
