+++ azuredevops/internal/service/securityroles/resource_securityrole_assignment.go
@@ -8,7 +8,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/client"
 	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils"
-	securityroles "github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils/clients"
+	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/utils/securityroles"
 )
 
 func ResourceSecurityRoleAssignment() *schema.Resource {
